+++ lib/samanage/error.rb
@@ -3,8 +3,14 @@ module Samanage
 	class Error < StandardError
 		attr_accessor :status_code, :response, :error
 		def initialize(error: nil, response: {})
-			self.status_code = response.nil? ? nil : response[:code]
-			self.response = response.nil? ? nil : response[:data] ||= response[:response]
+			if response.class == Hash
+				self.status_code = response[:code]
+				self.response = response[:data] ||= response[:response]
+			else
+				self.status_code = nil
+				self.response = response
+			end
+
 			self.error = error
 			puts "[Error] #{self.status_code}: #{self.response}"
 		end