+++ changelog.md
@@ -1,5 +1,18 @@
+#1.7.6
+- Adding group_id find methods for group name and user email
+
+#1.7.5
+- Adding site, department, group creation
+
+#1.7.4
+- Solving admin listing issue
+
+#1.7.2
+- Catching refused connections as Samanage::Error
+
 #1.7.1
 - Client Side SSL certificate Forced
+
 #1.7.0
 - Switched to HTTParty
 - Payload now responds to `Hash`. `Strings` will be parsed using `JSON.parse(payload)`